add - final version

+++ frontend/lib/features/daily_news/data/repository/article_repository_impl.dart
@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import 'package:dio/dio.dart';
 import 'package:news_app_clean_architecture/core/constants/constants.dart';
 import 'package:news_app_clean_architecture/features/daily_news/data/data_sources/local/app_database.dart';
@@ -12,8 +10,8 @@ import '../data_sources/remote/news_api_service.dart';
 
 class ArticleRepositoryImpl implements ArticleRepository {
   final NewsApiService _newsApiService;
-  final AppDatabase _appDatabase;
-  ArticleRepositoryImpl(this._newsApiService,this._appDatabase);
+  final AppDatabase? _appDatabase;
+  ArticleRepositoryImpl(this._newsApiService, [this._appDatabase]);
   
   @override
   Future<DataState<List<ArticleModel>>> getNewsArticles() async {
@@ -24,36 +22,40 @@ class ArticleRepositoryImpl implements ArticleRepository {
       category:categoryQuery,
     );
 
-    if (httpResponse.response.statusCode == HttpStatus.ok) {
+    if (httpResponse.response.statusCode == 200) {
       return DataSuccess(httpResponse.data);
     } else {
       return DataFailed(
-        DioError(
-          error: httpResponse.response.statusMessage,
-          response: httpResponse.response,
-          type: DioErrorType.response,
-          requestOptions: httpResponse.response.requestOptions
-        )
+        Exception(httpResponse.response.statusMessage),
       );
     }
-   } on DioError catch(e){
+   } on Exception catch(e){
     return DataFailed(e);
    }
   }
 
   @override
   Future<List<ArticleModel>> getSavedArticles() async {
+    if (_appDatabase == null) return [];
     return _appDatabase.articleDAO.getArticles();
   }
 
   @override
   Future<void> removeArticle(ArticleEntity article) {
+    if (_appDatabase == null) return Future.value();
     return _appDatabase.articleDAO.deleteArticle(ArticleModel.fromEntity(article));
   }
 
   @override
   Future<void> saveArticle(ArticleEntity article) {
+    if (_appDatabase == null) return Future.value();
     return _appDatabase.articleDAO.insertArticle(ArticleModel.fromEntity(article));
   }
   
+  // Custom Method para Technical Test - Mocked temporalmente mientras no conectamos Firebase
+  @override
+  Future<void> publishArticle(ArticleEntity article) async {
+    // Esto se re-escribirá en la fase 4 cuando conectemos firestore
+    return;
+  }
 }
